move Context to render layer
